Made Windows dependencies optional when installing on another platform
The pywin32 would break installation on a Linux as pip could not
find a suitable build, as the only binaries available are for win32.
So I made it made optional for others platforms.

I also made pyinstaller optional, does it make sense to include it in
other platforms ?
Is it used in another way that when building the .exe files ?
If yes, feel free to revert the change.