Bump Flask-SeaSurf to 2.0.0
Rather than to a specific GitHub commit.
As Flask-Seasurf on PyPi (v1.1.1) was broken with Flask 3+, and not
updated anymore, we had to point in the mtw_requirements.txt to
a specific commit on GitHub instead.
The commit was merged for the 2.0.0 release and doesn't exist anymore,
so pointing to the specific commit break the pip process now.